requirements.txt
django==4.2.13
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
groq==1.1.2
tiktoken==0.12.0
tenacity==8.3.0
streamlit
requests
pytest
pytest-django
pymupdf==1.27.2.2
chromadb==0.5.5
sentence-transformers==3.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-groq==0.1.9
